restore circular import
