+++ nilearn/image/image.py
@@ -535,3 +535,50 @@ def iter_img(imgs):
     output: iterator of 3D nibabel.Nifti1Image
     """
     return check_niimg_4d(imgs, return_iterator=True)
+
+
+def new_img_like(ref_niimg, data, affine=None, copy_header=False):
+    """Create a new image of the same class as the reference image
+
+    Parameters
+    ----------
+    ref_niimg: image
+        Reference image. The new image will be of the same type.
+
+    data: numpy array
+        Data to be stored in the image
+
+    affine: 4x4 numpy array
+        Transformation matrix
+
+    copy_header: boolean, optional
+        Indicated if the header of the reference image should be used to
+        create the new image
+
+    Returns
+    -------
+
+    new_img: image
+        An image which has the same type as the reference image.
+    """
+    from .._utils import load_niimg  # avoid circular import
+
+    ref_niimg = load_niimg(ref_niimg)
+
+    if affine is None:
+        affine = ref_niimg.get_affine()
+    if data.dtype == bool:
+        default_dtype = np.int8
+        if (LooseVersion(nibabel.__version__) >= LooseVersion('1.2.0') and
+                isinstance(ref_niimg, nibabel.freesurfer.mghformat.MGHImage)):
+            default_dtype = np.uint8
+        data = as_ndarray(data, dtype=default_dtype)
+    header = None
+    if copy_header:
+        header = copy.copy(ref_niimg.get_header())
+        header['scl_slope'] = 0.
+        header['scl_inter'] = 0.
+        header['glmax'] = 0.
+        header['cal_max'] = np.max(data) if data.size > 0 else 0.
+        header['cal_max'] = np.min(data) if data.size > 0 else 0.
+    return ref_niimg.__class__(data, affine, header=header)
