+++ nilearn/image/tests/data/test.mgz
